pyproject.toml
[tool.poetry]
name = "concurrent-scrape"
version = "0.1.0"
description = ""
authors = ["delphai <admin@delphai.com>"]

[tool.poetry.dependencies]
python = "^3.8"
delphai-utils = {extras = ["config","grpc"], version = "^0.2.0"}
tqdm = "^4.61.0"
pandas = "^1.2.4"

[tool.poetry.dev-dependencies]
yapf = "^0.30.0"
pep8-naming = "^0.11.1"
flake8 = "^3.8.3"
poethepoet = "^0.8.0"
grpcio-tools = "^1.32.0"
mypy = "^0.800"
mypy-protobuf = "^2.4"

[tool.poe.tasks]
start = "python src/main.py"
codegen = [
  {cmd = "mkdir -p ./src/proto"},
  {cmd = "protodep up"},
  {cmd = "python -m grpc_tools.protoc -I./.vendor -I. --mypy_out=./src/proto --grpc_python_out=./src/proto --python_out=./src/proto ./.vendor/**/*.proto"}
]

[tool.pytest.ini_options]
addopts = "-p no:warnings"
filterwarnings = [
  "error"
]